+++ components/ui/molecules/index.ts
@@ -1,0 +1,2 @@
+export { default as TabBar } from './TabBar'
+export { default as TabItem } from './TabItem'
